KKUMI-68 [FEAT] #61 - 포스트 작성 > 이미지에 핀 추가 / 이미지 선택 변경 없이 첫 핀 추가할 때 제대로 동작 안 하는 문제 해결 필요

+++ core/domain/src/main/java/com/swmarastro/mykkumi/domain/entity/PostEditVO.kt
@@ -4,8 +4,8 @@ package com.swmarastro.mykkumi.domain.entity
 
 data class PostEditPinVO(
     var pinId: Long?,
-    var positionX: Double,
-    var positionY: Double,
+    var positionX: Float,
+    var positionY: Float,
 //    var product: PostEditPinProductVO?
 )
